refactor(core): S4 disambiguate delivery/daemon twin namings (#428)
- daemon.ts (26KB reducer) → daemon/reduce.ts to end file-shadows-dir ambiguity
- daemon/delivery.ts (daemon-direct loop) → daemon/delivery-loop.ts to distinguish from delivery/ defer primitives
- Drop DeferDelivery re-export hop in workspaces/runtimes/cmux.ts; each throw-site now imports from @squadrant/core directly
- workspaces/cmux/ → workspaces/cmux-daemon/ so the daemon utility dir no longer collides by name with runtimes/cmux.ts

No behavior changes. gitnexus_detect_changes: 0 symbols changed. 1586/1586 tests green.
