+++ packages/workspaces/src/cmux-daemon/daemon-cmux.ts
@@ -1,5 +1,5 @@
 import type { RuntimeDriver, PaneRef } from "../runtimes/types.js";
-import { DeferDelivery } from "../runtimes/cmux.js";
+import { DeferDelivery } from "@squadrant/core";
 
 /**
  * #332: daemon-side cmux access. The daemon (a launchd process, NOT a cmux
