+++ packages/workspaces/src/index.ts
@@ -2,7 +2,7 @@
 export * from "./runtimes/index.js";
 export * from "./notifiers/index.js";
 export * from "./workspaces/index.js";
-export { CmuxEventsBridge, deriveRunState } from "./cmux/events-bridge.js";
-export type { RunState, CmuxEventsChild, CmuxAgentHook, CmuxEventsBridgeDeps } from "./cmux/events-bridge.js";
-export { DaemonCmux } from "./cmux/daemon-cmux.js";
+export { CmuxEventsBridge, deriveRunState } from "./cmux-daemon/events-bridge.js";
+export type { RunState, CmuxEventsChild, CmuxAgentHook, CmuxEventsBridgeDeps } from "./cmux-daemon/events-bridge.js";
+export { DaemonCmux } from "./cmux-daemon/daemon-cmux.js";
 export { getFreePort, listProjectCrews, findCrew, resolveCaptainWorkspace, sendFirstTurnWhenReady } from "./crew-pane.js";
